nl80211: fix the "not disabled" case in disable_11b_rates()
the NL80211_ATTR_TX_RATES attribute should always be part
of a NL80211_CMD_SET_TX_BITRATE_MASK command.

In case the disabled parameter in the nl80211_disable_11b_rates()
is false, the attribute that should be left out of the command
message is NL80211_TXRATE_LEGACY - this way, the bitrate mask
will be reset to its default.

Signed-off-by: Guy Eilam <guy@wizery.com>